Update options
This changes the available options to be more generic for use with non
iOS platforms. This replaces the previous `iOSsdkPath` with a more
generic `sdkPath` to make it clear this doesn't have to be iOS specific.
It also adds `sdk` which allows a SDK name so you don't have to hardcode
the specific path. If you don't pass either of these the default macOS
SDK is used. This also exposes arbitrary arguments for users to pass
other args.